+++ src/components/Nav/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from './Button';
+import Button from '../Button/Button.js';
 import './Nav.css'
 
 function Nav({setStage}) {
