+++ src/components/TripCard/TripCard.css
