+++ src/components/TripCard/TripCard.js
