+++ src/components/TripPage/TripPage.css
