+++ src/components/TripSection/TripSection.css
